# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Kimdasa Construction Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leads:
    get:
      operationId: getLeads
      tags: [leads]
      summary: List all leads
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create a new lead (public endpoint for website form)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeadRequest"
      responses:
        "201":
          description: Lead created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get a lead by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Not found
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update a lead
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLeadRequest"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete a lead
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /estimates:
    get:
      operationId: getEstimates
      tags: [estimates]
      summary: List all estimates
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of estimates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Estimate"
    post:
      operationId: createEstimate
      tags: [estimates]
      summary: Create an estimate
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEstimateRequest"
      responses:
        "201":
          description: Estimate created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Estimate"
  /estimates/calculate:
    post:
      operationId: calculateEstimate
      tags: [estimates]
      summary: Calculate estimate cost breakdown
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalculateEstimateRequest"
      responses:
        "200":
          description: Cost breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EstimateCalculation"
  /estimates/{id}:
    get:
      operationId: getEstimate
      tags: [estimates]
      summary: Get an estimate by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Estimate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Estimate"
    patch:
      operationId: updateEstimate
      tags: [estimates]
      summary: Update an estimate
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEstimateRequest"
      responses:
        "200":
          description: Updated estimate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Estimate"
    delete:
      operationId: deleteEstimate
      tags: [estimates]
      summary: Delete an estimate
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /follow-ups:
    get:
      operationId: getFollowUps
      tags: [follow-ups]
      summary: List follow-ups
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of follow-ups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FollowUp"
    post:
      operationId: createFollowUp
      tags: [follow-ups]
      summary: Create a follow-up
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFollowUpRequest"
      responses:
        "201":
          description: Follow-up created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowUp"
  /follow-ups/{id}:
    patch:
      operationId: updateFollowUp
      tags: [follow-ups]
      summary: Update a follow-up
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFollowUpRequest"
      responses:
        "200":
          description: Updated follow-up
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowUp"
    delete:
      operationId: deleteFollowUp
      tags: [follow-ups]
      summary: Delete a follow-up
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    Lead:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        serviceType:
          type: string
          nullable: true
        urgency:
          type: string
          nullable: true
        budgetApprox:
          type: string
          nullable: true
        bestContactTime:
          type: string
          nullable: true
        comments:
          type: string
          nullable: true
        status:
          type: string
          enum: [new_lead, contacted, estimate_sent, approved, in_progress, completed, lost]
        language:
          type: string
        source:
          type: string
          nullable: true
        assignedTo:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - status
        - language
        - createdAt
        - updatedAt
    CreateLeadRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        serviceType:
          type: string
        urgency:
          type: string
        budgetApprox:
          type: string
        bestContactTime:
          type: string
        comments:
          type: string
        status:
          type: string
          enum: [new_lead, contacted, estimate_sent, approved, in_progress, completed, lost]
        language:
          type: string
        source:
          type: string
        assignedTo:
          type: string
      required:
        - name
    UpdateLeadRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        serviceType:
          type: string
        urgency:
          type: string
        budgetApprox:
          type: string
        bestContactTime:
          type: string
        comments:
          type: string
        status:
          type: string
          enum: [new_lead, contacted, estimate_sent, approved, in_progress, completed, lost]
        language:
          type: string
        source:
          type: string
        assignedTo:
          type: string
...
    Estimate:
      type: object
      properties:
        id:
          type: integer
        leadId:
          type: integer
          nullable: true
        customerId:
          type: integer
          nullable: true
        serviceType:
          type: string
        measurements:
          type: string
          nullable: true
        difficulty:
          type: string
          nullable: true
        height:
          type: string
          nullable: true
        roofPitch:
          type: string
          nullable: true
        materials:
          type: string
          nullable: true
        region:
          type: string
          nullable: true
        laborCost:
          type: string
          nullable: true
        materialCost:
          type: string
          nullable: true
        overhead:
          type: string
          nullable: true
        profitMargin:
          type: string
          nullable: true
        internalCost:
          type: string
          nullable: true
        recommendedPrice:
          type: string
          nullable: true
        clientPrice:
          type: string
          nullable: true
        marginPercent:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, sent, approved, rejected]
        notes:
          type: string
          nullable: true
        internalNotes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - serviceType
        - status
        - createdAt
        - updatedAt
    CreateEstimateRequest:
      type: object
      properties:
        leadId:
          type: integer
        customerId:
          type: integer
        serviceType:
          type: string
        measurements:
          type: string
        difficulty:
          type: string
        height:
          type: string
        roofPitch:
          type: string
        materials:
          type: string
        region:
          type: string
        laborCost:
          type: string
        materialCost:
          type: string
        overhead:
          type: string
        profitMargin:
          type: string
        internalCost:
          type: string
        recommendedPrice:
          type: string
        clientPrice:
          type: string
        marginPercent:
          type: string
        status:
          type: string
          enum: [draft, sent, approved, rejected]
        notes:
          type: string
        internalNotes:
          type: string
      required:
        - serviceType
    CalculateEstimateRequest:
      type: object
      properties:
        serviceType:
          type: string
        squareFootage:
          type: number
        difficulty:
          type: string
          enum: [easy, medium, hard]
        region:
          type: string
          enum: [new_jersey, pennsylvania]
      required:
        - serviceType
    EstimateCalculation:
      type: object
      properties:
        serviceType:
          type: string
        squareFootage:
          type: number
        difficulty:
          type: string
        region:
          type: string
        materialCost:
          type: string
        laborCost:
          type: string
        overhead:
          type: string
        profitMargin:
          type: string
        internalCost:
          type: string
        recommendedPrice:
          type: string
        marginPercent:
          type: string
        basePriceUsed:
          nullable: true
          type: object
          properties:
            category:
              type: string
            avgPrice:
              type: string
      required:
        - serviceType
        - materialCost
        - laborCost
        - overhead
        - profitMargin
        - internalCost
        - recommendedPrice
        - marginPercent
...
    FollowUp:
      type: object
      properties:
        id:
          type: integer
        leadId:
          type: integer
          nullable: true
        customerId:
          type: integer
          nullable: true
        scheduledAt:
          type: string
          format: date-time
        notes:
          type: string
          nullable: true
        completed:
          type: boolean
        completedAt:
          type: string
          format: date-time
          nullable: true
        assignedTo:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - scheduledAt
        - completed
        - createdAt
        - updatedAt
    CreateFollowUpRequest:
      type: object
      properties:
        leadId:
          type: integer
        customerId:
          type: integer
        scheduledAt:
          type: string
          format: date-time
        notes:
          type: string
        completed:
          type: boolean
        completedAt:
          type: string
          format: date-time
        assignedTo:
          type: string
      required:
        - scheduledAt
...